+++ CHANGELOG.md
@@ -3,16 +3,19 @@
 ## 0.9.1
 
 ### Added
+
 - `Transform(Plane plane)`
 - `Polygon.Trimmed(Plane plane, bool flip)`
 - `GetSolid()` method on GeometricElement that returns the Csg solid.
 - `Polygon.ToTransform()`
 - `Elements.Anaysis.AnalysisImage`
+- `Profile.CreateFromPolygons(IEnumerable<Polygon> polygons)`
 
 ### Changed
+
 - `AnalysisMesh` now handles single valued analysis.
 - `Polygon.Split()` can now handle polygons that are not in the XY plane.
-- Leave the discriminator property during deserialization.  It will go to AdditionalProperties.
+- Leave the discriminator property during deserialization. It will go to AdditionalProperties.
 
 ## 0.9.0